Last day rush: userland (unstable)
pre merge programs branch

+++ Kernel/include/time.h
@@ -1,8 +1,9 @@
 #ifndef _TIME_H_
 #define _TIME_H_
+#include <stdint.h>
 
 void timer_handler();
-int ticks_elapsed();
-int seconds_elapsed();
+uint64_t ticks_elapsed();
+uint64_t seconds_elapsed();
 
 #endif
